+++ zerver/lib/topic.py
@@ -5,8 +5,6 @@ import orjson
 from django.db import connection
 from django.db.models import F, Func, JSONField, Q, QuerySet, Subquery, TextField, Value
 from django.db.models.functions import Cast
-from sqlalchemy.sql import ColumnElement, column, func, literal
-from sqlalchemy.types import Boolean, Text
 
 from zerver.lib.request import REQ
 from zerver.lib.types import EditHistoryEvent
@@ -70,22 +68,6 @@ using "subject" in the DB sense, and nothing customer facing.
 # zerver/lib/message.py, and it's not user facing.
 DB_TOPIC_NAME = "subject"
 MESSAGE__TOPIC = "message__subject"
-
-
-def topic_match_sa(topic_name: str) -> ColumnElement[Boolean]:
-    # _sa is short for SQLAlchemy, which we use mostly for
-    # queries that search messages
-    topic_cond = func.upper(column("subject", Text)) == func.upper(literal(topic_name))
-    return topic_cond
-
-
-def get_resolved_topic_condition_sa() -> ColumnElement[Boolean]:
-    resolved_topic_cond = column("subject", Text).startswith(RESOLVED_TOPIC_PREFIX)
-    return resolved_topic_cond
-
-
-def topic_column_sa() -> ColumnElement[Text]:
-    return column("subject", Text)
 
 
 def filter_by_topic_name_via_message(
